[196] Extract deresolving name logic so it can be used during import
Bug: https://github.com/eclipse-syson/syson/issues/196
